Move generic widget to better place
